pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Tobias Waslowski <tobiaswaslowski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-iso639 = "^2024.1.2"
shared = {git = "https://github.com/twaslowski/lingolift-shared"}

[tool.poetry.group.generative.dependencies]
openai = "^1.7.2"
json5 = "^0.9.14"

# spacy dependencies
[tool.poetry.group.nlp.dependencies]
spacy = "^3.7.0"
de-core-news-sm = { url = "https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0.tar.gz" }
es-core-news-sm = { url = "https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0.tar.gz" }
ru-core-news-sm = { url = "https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.7.0/ru_core_news_sm-3.7.0.tar.gz" }
lingua-language-detector = "^2.0.2"


# testing, linting, utilities
[tool.poetry.group.dev.dependencies]
pip = "^23.3.2"
mypy = "^1.8.0"
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
poetry-plugin-export = "^1.5.0"
urllib3 = "<2"
black = "^24.4.2"
isort = "^5.13.2"

[tool.poetry.group.webserver.dependencies]
flask = "^3.0.1"
flask-cors = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.4.4"
addopts = "-s -v --disable-warnings --ignore=test/integration/"
testpaths = [
    "test",
]